feat(router): add safety pipeline, structured telemetry, and fix cosine precision
- safety pipeline: add prompt-injection detection filter
  (`filters/injection_detect.rs`) with weighted regex rules, score
  threshold, category deduplication, and 12 tests.  Add secret-masking
  transform (`transforms/secret_mask.rs`) for Bearer/basic auth, API keys,
  AWS/GitHub tokens, preserving key prefixes.  Add output- sanitization
  transform (`transforms/sanitize.rs`) removing bidi overrides, C1 control
  chars, Plane-14 tags, and ANSI escape sequences.  Wire injection filter
  into DeterministicPreFilter's builtin filter list.  38 new tests total.

- structured telemetry: add `FailureClass` closed enum (8 variants)
  replacing free-string error labels in `RouterMetrics`; `classify_error()`
  regex-based auto-classifier with fast-path for compiler diagnostic codes;
  `TelemetryEvent` enum-gated events with controlled-vocabulary fields (no
  free strings, no PII); `TelemetrySink` trait with `TracingSink` and
  `NoopSink` implementations; optional sink wired into
  `RouterMetrics::record_error`.  29 new tests.

- search-vector precision fix: replace `as f64` casts with `f64::from()` in
  cosine similarity accumulators to satisfy clippy::cast_lossless.

- token utilities: replace flat byte-count token estimation
  with Unicode-script-range character weighting in `common-core::tokens`.
  CJK, emoji, whitespace, and control chars each get distinct per-char
  token weights. Add `estimate_tokens_floor` clamped-to-minimum variant.
  Deprecate `DEFAULT_CHARS_PER_TOKEN`. Add `AtomicTokenBudget` — a
  thread-safe token budget using `AtomicU64` CAS loops, supporting
  reserve/release/reset with `&self` methods. Add `chunk_document` —
  sentence-boundary-aware document chunker with CJK/full-width
  terminators, newline boundaries, configurable max tokens, and overlap
  preservation. 20 tests.

- cosine precision: change dot-product and norm
  accumulators in `cosine_similarity` from f32 to f64 for numerically
  stable ranking in high-dimensional embedding spaces. f64::from() casts
  satisfy clippy::cast_lossless. Add two regression tests proving correct
  ranking order at 384 and 512 dimensions.

+++ src/job-copilot/src/prompt/context.rs
@@ -13,7 +13,7 @@ pub fn select_lod(page_context: &str, token_budget: usize) -> (usize, &'static s
     // Levels 0–4 map to indices 0–4; skip level 5 (alias of full text).
     let limit = 5.min(slices.len());
     for (level, slice) in slices.iter().take(limit).enumerate() {
-        if estimate_tokens(slice) <= token_budget {
+        if estimate_tokens(slice) <= token_budget as u64 {
             return (level, LOD_LABELS[level]);
         }
     }
@@ -33,7 +33,7 @@ pub fn chunk_page_context(page_context: &str, token_budget: usize) -> String {
     // Find best level (0–4).
     let limit = 5.min(slices.len());
     for slice in slices.iter().take(limit) {
-        if estimate_tokens(slice) <= token_budget {
+        if estimate_tokens(slice) <= token_budget as u64 {
             return slice.clone();
         }
     }
